law_chatbot/requirements.txt
fastapi
uvicorn
ollama
streamlit
faiss-cpu
transformers
torch
PyPDF2
requests 